Refactor API code
The code has been refactored with multiple changes:
- Tokens' endpoints separated using routers.
- Auxiliary files separated into modules.
- Tests updated for the new organization.
- Programmer mode removed, only agent allowed now for better code
  maintenance. Also, current open source LLMs are already good for agent
  tasks.
- Pre-commit now includes unused import automatic removal.
