+++ .pre-commit-config.yaml
@@ -18,6 +18,12 @@ repos:
       - id: check-added-large-files
         exclude: poetry.lock
 
+  - repo: https://github.com/hadialqattan/pycln
+    rev: v2.4.0
+    hooks:
+      - id: pycln
+        args: [--config=pyproject.toml]
+
   # python code formatting
   - repo: https://github.com/psf/black
     rev: 23.1.0
