+++ gptstonks_api/agent/__init__.py
@@ -1,0 +1,1 @@
+from .run_background import run_agent_in_background as run_agent_in_background
