+++ gptstonks_api/constants/__init__.py
@@ -1,0 +1,2 @@
+from .constants import AI_PREFIX as AI_PREFIX
+from .constants import API_DESCRIPTION as API_DESCRIPTION
