+++ gptstonks_api/databases/__init__.py
@@ -1,0 +1,1 @@
+from .mongodb import db as db
