+++ gptstonks_api/initialization/__init__.py
@@ -1,0 +1,1 @@
+from .app import init_api as init_api
