--- gptstonks_api/models.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class TokenData(BaseModel):
-    openbb: str
