+++ gptstonks_api/utils/__init__.py
@@ -1,0 +1,2 @@
+from .openbb_chat_qa import get_openbb_chat_output as get_openbb_chat_output
+from .openbb_chat_qa import run_repl_over_openbb as run_repl_over_openbb
