+++ tests/test_multiple_pats.py
@@ -3,7 +3,9 @@ import asyncio
 import pytest
 from httpx import AsyncClient, codes
 
-from gptstonks_api.main import app, init_data
+from gptstonks_api.databases import db
+from gptstonks_api.initialization import init_api
+from gptstonks_api.main import app, app_data
 
 
 async def run_multiple_async_queries():
@@ -14,7 +16,6 @@ async def run_multiple_async_queries():
                     "/process_query_async",
                     json={
                         "query": "get the balance sheets for AAPL",
-                        "use_agent": True,
                         "openbb_pat": str(i),
                     },
                 )
@@ -27,10 +28,10 @@ async def run_multiple_async_queries():
 @pytest.mark.asyncio
 async def test_multiple_pats():
     try:
-        init_data()
-    except AttributeError as e:
-        pytest.skip("No .env file provided, or not all env variables specified")
-
+        db.command("ping")
+    except Exception as e:
+        pytest.skip("No database to connect to. Test skipped")
+    init_api(app_data)
     res = await run_multiple_async_queries()
     for r in res:
         assert r.status_code == codes.OK
